+++ src/components/layouts/AdminLayout.tsx
@@ -1,0 +1,12 @@
+import React, { PropsWithChildren } from "react";
+
+const AdminLayout = ({ children }: PropsWithChildren) => {
+    return (
+        <>
+            <p>Admin</p>
+            {children}
+        </>
+    )
+}
+
+export default AdminLayout
