+++ src/components/layouts/MainLayout/index.tsx
@@ -1,0 +1,15 @@
+import React, { PropsWithChildren } from 'react';
+import Header from './Header'
+import Footer from './Footer'
+
+const MainLayout = ({ children }: PropsWithChildren) => {
+    return (
+        <>
+            <Header />
+            {children}
+            <Footer />
+        </>
+    )
+}
+
+export default MainLayout
